+++ google-cloud/container-linux/kubernetes/controllers/controllers.tf
@@ -14,12 +14,18 @@ resource "google_dns_record_set" "etcds" {
   rrdatas = ["${element(google_compute_instance.controllers.*.network_interface.0.address, count.index)}"]
 }
 
+# Zones in the region
+data "google_compute_zones" "available" {
+  region = "${var.region}"
+  status = "UP"
+}
+
 # Controller instances
 resource "google_compute_instance" "controllers" {
   count = "${var.count}"
 
   name         = "${var.cluster_name}-controller-${count.index}"
-  zone         = "${var.zone}"
+  zone         = "${element(data.google_compute_zones.available.names, count.index)}"
   machine_type = "${var.machine_type}"
 
   metadata {