+++ CHANGES.md
@@ -9,9 +9,11 @@ Notable changes between versions.
 * Add required variable `region` (e.g. "us-central1")
 * Reduce time to bootstrap a cluster
 * Change etcd to run on-host, across controllers (etcd-member.service)
-* Change worker managed instance group to automatically span zones in a region
+* Change controller instances to automatically span zones in the region
+* Change worker managed instance group to automatically span zones in the region
 * Remove support for self-hosted etcd
-* Remove `controller_preemptible` optional variable
+* Remove the `zone` required variable
+* Remove the `controller_preemptible` optional variable
 
 ## v1.8.2